requirements.txt
accelerate==1.1.1
annotated-types==0.7.0
anyio==4.6.2.post1
appscript==1.3.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
docx2pdf==0.1.8
fastapi==0.115.5
filelock==3.16.1
fonttools==4.55.0
fsspec==2024.10.0
h11==0.14.0
huggingface-hub==0.26.3
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
psutil==6.1.0
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
pyparsing==3.2.0
PyPDF2==3.0.1
pypdfium2==4.30.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-multipart==0.0.18
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.13.3
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.67.1
transformers==4.31.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
